pyproject.toml
[project]
name = "rd-to-bq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "db-dtypes>=1.4.3",
    "google-cloud-bigquery>=3.34.0",
    "pandas>=2.3.1",
    "pandas-gbq>=0.29.2",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
